Bring back up to date

+++ NBXplorer.Tests/ServerTester.cs
@@ -22,6 +22,7 @@ using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Configuration;
 using NBXplorer.Logging;
 using NBXplorer.DerivationStrategy;
+using System.Net.Http;
 
 namespace NBXplorer.Tests
 {
@@ -59,9 +60,11 @@ namespace NBXplorer.Tests
 			get; set;
 		}
 
+		public string Caller { get; }
 		public ServerTester(string directory, bool autoStart = true)
 		{
 			SetEnvironment();
+			Caller = directory;
 			var rootTestData = "TestData";
 			directory = Path.Combine(rootTestData, directory);
 			_Directory = directory;
@@ -71,13 +74,33 @@ namespace NBXplorer.Tests
 				Start();
 		}
 
+		public async Task Load(string dataName)
+		{
+			datadir = Path.Combine(_Directory, "explorer");
+			if (Directory.Exists(datadir))
+				DeleteFolderRecursive(datadir);
+			Directory.CreateDirectory(_Directory);
+			Directory.CreateDirectory(datadir);
+			datadir = Path.Combine(datadir, "RegTest", "db");
+			Directory.CreateDirectory(datadir);
+			foreach (var file in Directory.GetFiles(Path.Combine("Data", dataName)))
+			{
+				File.Copy(file, Path.Combine(datadir, Path.GetFileName(file)));
+			}
+			LoadedData = true;
+			await using var db = await DBTrie.DBTrieEngine.OpenFromFolder(datadir);
+			using var tx = await db.OpenTransaction();
+			await tx.GetTable("IndexProgress").Delete();
+		}
+
 		public void Start()
 		{
 			try
 			{
 				var cryptoSettings = new NBXplorerNetworkProvider(NetworkType.Regtest).GetFromCryptoCode(CryptoCode);
 				NodeBuilder = NodeBuilder.Create(nodeDownloadData, Network, _Directory);
-
+				if (KeepPreviousData)
+					NodeBuilder.CleanBeforeStartingNode = false;
 				Explorer = NodeBuilder.CreateNode();
 				Explorer.ConfigParameters.Add("txindex", "1");
 				foreach (var node in NodeBuilder.Nodes)
@@ -86,10 +109,12 @@ namespace NBXplorer.Tests
 					node.CookieAuth = cryptoSettings.SupportCookieAuthentication;
 				}
 				NodeBuilder.StartAll();
-				Explorer.CreateRPCClient().EnsureGenerate(Network.Consensus.CoinbaseMaturity + 1);
+				if (!KeepPreviousData)
+					Explorer.CreateRPCClient().EnsureGenerate(Network.Consensus.CoinbaseMaturity + 1);
 
 				datadir = Path.Combine(_Directory, "explorer");
-				DeleteFolderRecursive(datadir);
+				if (!KeepPreviousData && !LoadedData)
+					DeleteFolderRecursive(datadir);
 				StartNBXplorer();
 				this.Client.WaitServerStarted();
 			}
@@ -100,6 +125,7 @@ namespace NBXplorer.Tests
 			}
 		}
 
+		public int TrimEvents { get; set; } = -1;
 		public bool UseRabbitMQ { get; set; } = false;
 		private void StartNBXplorer()
 		{
@@ -109,12 +135,15 @@ namespace NBXplorer.Tests
 			keyValues.Add(("datadir", datadir));
 			keyValues.Add(("port", port.ToString()));
 			keyValues.Add(("network", "regtest"));
+			keyValues.Add(("instancename", Caller));
 			keyValues.Add(("chains", CryptoCode.ToLowerInvariant()));
 			keyValues.Add(("verbose", "1"));
 			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcauth", Explorer.GetRPCAuth()));
 			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}rpcurl", Explorer.CreateRPCClient().Address.AbsoluteUri));
 			keyValues.Add(("cachechain", "0"));
+			keyValues.Add(("exposerpc", "1"));
 			keyValues.Add(("rpcnotest", "1"));
+			keyValues.Add(("trimevents", TrimEvents.ToString()));
 			keyValues.Add(("mingapsize", "3"));
 			keyValues.Add(("maxgapsize", "8"));
 			keyValues.Add(($"{CryptoCode.ToLowerInvariant()}startheight", Explorer.CreateRPCClient().GetBlockCount().ToString()));
@@ -156,17 +185,23 @@ namespace NBXplorer.Tests
 			Host.Start();
 			Configuration = conf;
 			_Client = NBXplorerNetwork.CreateExplorerClient(Address);
+			HttpClient = ((IHttpClientFactory)Host.Services.GetService(typeof(IHttpClientFactory))).CreateClient();
+			HttpClient.BaseAddress = Address;
 			_Client.SetCookieAuth(Path.Combine(conf.DataDir, ".cookie"));
 			Notifications = _Client.CreateLongPollingNotificationSession();
 		}
 
+		public HttpClient HttpClient { get; internal set; }
+
 		string datadir;
 
-		public void ResetExplorer()
+		public void ResetExplorer(bool deleteAll = true)
 		{
 			Host.Dispose();
-			DeleteFolderRecursive(datadir);
+			if (deleteAll)
+				DeleteFolderRecursive(datadir);
 			StartNBXplorer();
+			this.Client.WaitServerStarted();
 		}
 
 		public LongPollingNotificationSession Notifications { get; set; }
@@ -342,16 +377,26 @@ namespace NBXplorer.Tests
 		}
 		public DerivationStrategyBase CreateDerivationStrategy(ExtPubKey pubKey, bool p2sh)
 		{
-			pubKey = pubKey ?? new ExtKey().Neuter();
+			key = key ?? new ExtKey();
+			pubKey = pubKey ?? key.Neuter();
 			string suffix = this.RPC.Capabilities.SupportSegwit ? "" : "-[legacy]";
 			suffix += p2sh ? "-[p2sh]" : "";
+			scriptPubKeyType = p2sh ? ScriptPubKeyType.SegwitP2SH : ScriptPubKeyType.Segwit;
 			return NBXplorerNetwork.DerivationStrategyFactory.Parse($"{pubKey.ToString(this.Network)}{suffix}");
+		}
+		ExtKey key;
+		ScriptPubKeyType scriptPubKeyType;
+		public void SignPSBT(PSBT psbt)
+		{
+			psbt.SignAll(key.AsHDScriptPubKey(scriptPubKeyType), key);
 		}
 
 		public bool RPCStringAmount
 		{
 			get; set;
 		} = true;
+		public bool KeepPreviousData { get; set; }
+		public bool LoadedData { get; private set; }
 
 		public uint256 SendToAddress(BitcoinAddress address, Money amount)
 		{
